fix(timelock-queue): home classifyBlockedRow where tests can reach it without typechain
The executor transitively imports demoScriptHelpers, which imports the generated
`typechain/` dir. The run-ts-tests CI job does not generate it (that is gated
behind the requires-types label), so a test file importing the executor fails in
CI while passing locally.

Move the pure classifier to timelock-queue.ts — its natural home as part of the
queue state machine, and dependency-light — and merge its tests into that
module's suite. Verified by running the whole TS suite with `typechain/` moved
aside, reproducing the CI condition: 828 pass.
