+++ script/deploy/safe/execute-pending-timelock-tx.ts
@@ -47,6 +47,7 @@ import {
 import { formatTimelockScheduleBatch } from './safe-decode-utils'
 import {
   byOperationId,
+  classifyBlockedRow,
   computeOperationIdBatch,
   deserializeScheduleParams,
   getTimelockQueueCollection,
@@ -618,30 +619,6 @@ async function countQueuedAndBlockedOps(networkName: string): Promise<{
   } finally {
     await client.close()
   }
-}
-
-/**
- * Classifies a blocked row from its freshly-read on-chain state.
- *
- * - `done` — the controller executed it; reconcile the row to `executed`.
- * - `gone` — the controller has no timestamp for the id, so it was cancelled (or
- *   never scheduled); reconcile the row to `cancelled`. Cancelling is the
- *   documented remediation for a blocked op, so this is the common path, and
- *   without it following that advice leaves a permanently misleading row.
- * - `pending` — still a live operation; a candidate for the ready-check alert.
- *
- * @param state - On-chain flags for the operation.
- * @returns Which reconciliation (if any) the row needs.
- */
-export function classifyBlockedRow(state: {
-  isDone: boolean
-  isPending: boolean
-  isReady: boolean
-  isOperation: boolean
-}): 'done' | 'gone' | 'pending' {
-  if (state.isDone) return 'done'
-  if (!state.isPending && !state.isReady && !state.isOperation) return 'gone'
-  return 'pending'
 }
 
 /**
